imgproc: sync boundingRect function with 5.x
